Update introduction-to-applicationhostconfig.md
Small documentation fix. Changed "The location of the file is currently in the system32\inetsrv directory, but this is expected to change after beta2 to system32\inetsrv\config." to "The location of the file is currently in the `%windir%\system32\inetsrv\config` directory." as this is where the file sits now.