Update description for the "Cookie" field (siteDefaults) (#154)
* Modified the description for the "Cookie" field

Tests of the Cookie field ("cs(Cookie)") in the logs show that only inbound cookies from HTTP requests are logged, not outbound cookies. This also makes sense considering the name of the field being "cs" for client->server.

* Update description for the "Cookie" field (site)

Tests of the Cookie field ("cs(Cookie)") in the logs show that only inbound cookies from HTTP requests are logged, not outbound cookies. This also makes sense considering the name of the field being "cs" for client->server. This PR is for the per-site logging settings.