minBytesPerSecond explanation changes
As per OS bug 18419530 - the way in which the minBytesPerSecond value is used in http.sys is the following:
- the driver will look at the number of bytes that it has to send across the wire
- it will divide this number by the value of minBytesPerSecond
- the resulting number, in seconds, will be the time needed to stream the entire response across the wire at a speed of minBytesPerSecond.
Hence: a file that is 2.1 MB in size (2048411 bytes) will take 8537 seconds to be streamed across at minBytesPerSecond of 240 bytes / second. This means that the http.sys driver will only close the connection some 2.37 hours after the download has begun, and not if the transfer rate goes below 240 bytes per second for only a period of time during the download.

This can be very misleading since it does not allow to defend against DOS attacks very effectively.

Fix formatting