Update paragraph "using server variables in rewrite rules"
What is described in paragraph "using server variables in rewrite rules" is not correct. The information In the list about http headers is not correct. A dash in the http header is not converted to an underscore. It should be stated that de underscore from the condition input is converted to a dash. When the statement from the text was true, I could handle a HTTP header with an underscore in it.