Updated DotNetNuke Documentation References
Given that the DotNetNuke installation environment that is supported with deployment through the Web Platform Installer or the Web Matrix installation is an older, not-current, installation version this update makes two notes to this documentation.

1. References added to the DNN Community site which is the proper documentation for all DNN Platform (Open Source) information
2. Added a callout/quote note to draw attention to the need to upgrade after installation to the latest version of DNN, with an included link to the instructions on that upgrade process.

If this is not a proper method to make this recommended change, please let me know.  I just wanted to make sure it was noted that the installation versions as outlined in this guide SHOULD NOT be used in production without upgrades.